Initial commit
In-progress skeleton of a simulation for the design and analysis of
navigation and control algorithms for a two-wheeled rover. I'm
attempting to learn best practices for organizing a Python package
without relying on Pycharm, so there are some empty files that need to
be filled out, e.g. __init__.py.